requirements.txt
PyQt5==5.15.10
requests==2.31.0
pillow==10.2.0